requirements.txt
qhue
zeroconf
PyYAML
requests
aiohttp
backoff